Removed non-pure fetch-related functions from fetchHandlers index.js.  So that they are not loaded by handlerMap.
