+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "traec",
-  "version": "1.1.69",
+  "version": "1.1.70",
   "description": "",
   "main": "dist/index.js",
   "dependencies": {
