+++ src/fetchHandlers/tenant/index.js
@@ -92,3 +92,23 @@ export const deleteSubTenant = ({ tenantId }) => {
   };
   return { fetchParams, stateParams: { stateSetFunc } };
 };
+
+/* ACTION DISPATCH */
+
+export const postTenantDispatch = ({}) => {
+  const fetchParams = {
+    method: "POST",
+    url: `/api/tenant/admin/dispatch/`,
+    apiId: "api_tenant_admin_dispatch_create",
+    requiredParams: [],
+    queryParams: {}
+  };
+  const stateSetFunc = (state, action) => {
+    const data = action.payload;
+    return state;
+  };
+  return {
+    fetchParams,
+    stateParams: { stateSetFunc }
+  };
+};
